moved globalStyles to constants folder to follow proper react-native procedure and update Colors import in app pages
