+++ app/(tabs)/expiration.tsx
@@ -2,7 +2,7 @@ import { StyleSheet } from 'react-native';
 
 import EditScreenInfo from '@/components/EditScreenInfo';
 import { Text, View } from '@/components/Themed';
-import { Colors } from './globalStyles';
+import { Colors } from '@/constants/globalStyles';
 
 export default function ExpirationTabScreen() {
   return (
