--- app/(tabs)/globalStyles.tsx
@@ -1,11 +1,0 @@
-export const Colors = {
-    mainText: '#371B34',
-    secondaryText: '#A598A4',
-    background: '#FFFFFF',
-
-    primaryCard: '#D2EBDD',
-    secondaryCard: '#EBF6F0',
-
-    navigationActive: '#371B34',
-    navigationInactive: '#CDD0E3',
-};
