+++ constants/globalStyles.ts
@@ -1,0 +1,12 @@
+export const Colors: { [key: string]: string } = {
+    mainText: '#371B34',
+    secondaryText: '#A8A8B8',
+    highlightText: '#9DA5C1',
+    background: '#FFFFFF',
+
+    primaryCard: '#E8F5ED',
+    secondaryCard: '#FCF9F9',
+
+    navigationActive: '#371B34',
+    navigationInactive: '#CDD0E3',
+};
