+++ constants/templateColors.ts
